CDH-66950. Relax event checks in RpcEnvSuite.
Because of network issues the events may not exactly match the lists
the test expects; so check that the needed events are on the list,
but allow unexpected events to also be there.

This is similar to what the test does on the 2.x branches, although
the test code there has been refactored.

(cherry picked from commit 8df1ab185bf6cf2429edce8c05bf1fa31e8ab107)