[SPARK-23941][MESOS] Mesos task failed on specific spark app name
Shell escaped the name passed to spark-submit and change how conf attributes are shell escaped.

This test has been tested manually with Hive-on-spark with mesos or with the use case described in the issue with the sparkPi application with a custom name which contains illegal shell characters.

With this PR, hive-on-spark on mesos works like a charm with hive 3.0.0-SNAPSHOT.

I state that this contribution is my original work and that I license the work to the project under the project’s open source license

Author: Bounkong Khamphousone <bounkong.khamphousone@ebiznext.com>

Closes #21014 from tiboun/fix/SPARK-23941.

(cherry picked from commit 6782359a04356e4cde32940861bf2410ef37f445)
Signed-off-by: Marcelo Vanzin <vanzin@cloudera.com>
(cherry picked from commit 682f05d638a67b869d0e170f6c99683445e2449d)