CDH-70114. [==C5_APPROVED_BUGFIX==] Decrypt input stream before uncompressing during shuffle phase
We were encrypting the shuffle data that is spilled to disk during the shuffle stage in DiskBlockObjectWriter but not decrypting the data while reading. This fix now decrypts the data before uncompressing it.

Testing Done:
Ran SortSuite."basic sorting using ExternalSort" after setting spark.shuffle.encryption to true in TestSQLContext. The test failed without the fix but runs well after that.