CDH-69018. Add Statistics for Erasure Coding to org.apache.spark.executor.TaskMetrics
This change adds bytesReadEC (erasure-coded bytes read) to
InputMetrics within TaskMetrics. It obtains this value from
the HDFS Filesystem object. This value is not plumbed into
sinks, it is only directly accessible via a TaskMetrics/InputMetrics
object. This code will remain downstream only for now.

Tested with a simple unit test that verifies the ability
to obtain the erasure coded metrics from the filesystem.

Change-Id: If080090b1fecf070b718e6a8866df0f5bc85bbe1