[spark] CDH-57150. Exit spark-shell/spark-submit/pyspark with correct error message if no client configuration found
Spark Applications fail with NoClassDefFound exception when spark applications are launched from a host which does not have client configurations (Spark Gateway role).

This fix adds a error message and exits application if it fails to find SPARK_DIST_CLASSPATH configured in the environment variable.

Testing Done:
Reproduced the issue by adding a new host to an existing cluster and ensured that the fix adds the error message. Also ensured that this error message is not printed when client configurations are available.

Change-Id: I65f3f3a0ed8218095871b37c13abd2455f2fcfd5