CDH-73607. Use reflection to remove circular dependency
Spark-core depended on hive for managing hive delegation tokens.  But
hive depends on spark, for hive-on-spark.  This led to a circular
dependency, and cauldron refused to build CDH.

As the dependency was small, the fix used here is just to use reflection
in spark-core to call hive, and remove the explicit dependency.

Change-Id: Ia4b67dad1c4947e62ec776ff5934ccce14d5cd97