CDH-74083. Remove native ORC data source from CDH.
The underlying issue is that the Hive code in CDH has a forked
version of ORC that is much older than the version Spark needs.
Since CDH does not support ORC, it's easier to remove that feature
than to try to make it work with Hive's ORC.

ORC support is available (but still unsupported) through the Hive
data source implementation.

The change removes the native ORC data source and just leaves enough
code to enable the Hive-side unit tests. It also adjusts a whole
bunch of unit tests to not test the native source.

Some unrelated changes to get unit tests going:

- restrict the Hive versions under test to the only supported version.
- adjust ORC filter syntax in tests (see CDH-56492).
- disable some low-hanging-fruit tests that we don't want in CDH.

Change-Id: Ia830313129bfca3ea5b35863ea91d47df71cef76