CDH-74147. Turn on class loader isolation in VersionsSuite.
I can't quite explain why the test was failing without it; all the needed
classes are in the classpath, so the class loader should be able to see them.
But still it was failing to load some Hive classes, and didn't seem to be
throwing a proper error about it to boot.

Turning on isolation and giving the loader an explicit list of the jars in the
classpath seems to fix the issue.

Change-Id: If98c0e21cf4cd7cfd1dc275ee7c1f80904d079f9