CDH-74189. Disable SparkLauncherSuite.testInProcessLauncher.
There's still some flakiness in this test; it's also been reported
in SPARK-23020 after the latest fixes. It does seem very rare though,
and I think I spotted what is wrong in the code.

The fix might not be simple, so for now disable this test to avoid
spurious failures in jenkins.

Change-Id: Ic02def300f3465350320d367edbbfea0cc467fc5