[SPARK-24601] Update Jackson to 2.9.6 (without updating jackson version for CLOUDERA-BUILD in 6.1)
Hi all,

Jackson is incompatible with upstream versions, therefore bump the Jackson version to a more recent one. I bumped into some issues with Azure CosmosDB that is using a more recent version of Jackson. This can be fixed by adding exclusions and then it works without any issues. So no breaking changes in the API's.

I would also consider bumping the version of Jackson in Spark. I would suggest to keep up to date with the dependencies, since in the future this issue will pop up more frequently.

Bump Jackson to 2.9.6

Compiled and tested it locally to see if anything broke.

Please review http://spark.apache.org/contributing.html before opening a pull request.

Closes #21596 from Fokko/fd-bump-jackson.

Authored-by: Fokko Driesprong <fokkodriesprong@godatadriven.com>
Signed-off-by: hyukjinkwon <gurwls223@apache.org>

CLOUDERA-BUILD Backport Notes. In backporting this commit to 6.1 we have ignored the parts that actually update the jackson version. This change contains only fixes that are required for having a successfull backport of SPARK-23429 to 6.1. Without this change there will be issues with parsing jsons especially in HistoryServerSuite and AppStatusListenerSuite.