CDH-74187. Correct database location for namenode HA
If hdfs HA is turned on after a hive database is already created, the db
location may still reference just one namenode, instead of the
nameservice, if users do not properly follow all upgrade instructions.
After this change, spark detects the misconfiguration and tries to
auto-adjust for it, since this is the behavior from hive as well.

This serves the same purpose as CDH-58542, just expands it to cover
another spot in the code introduced by SPARK-19724.

Change-Id: I99f9a3575dc325c18cfd07676fa5eb035f0bb487