CDH-74725. Disable timeout test in BarrierTaskContextSuite.
The test seems inherently flaky because the timeout starts when
the first task syncs; so if the first task takes a little bit of
time to sync, that compensates for the sleep time in the "slow"
task.

Increasing the sleep could help, but it also increases the run
time of the test.

Change-Id: I9a91cdf7caa310ee5f7fe4d3c8f6330245915b7c
(cherry picked from commit b820994f682f7daf1317eb446ba5020e17a12969)