[SPARK-26003] Improve SQLAppStatusListener.aggregateMetrics performance
In `SQLAppStatusListener.aggregateMetrics`, we use the `metricIds` only to filter the relevant metrics. And this is a Seq which is also sorted. When there are many metrics involved, this can be pretty inefficient. The PR proposes to use a Set for it.

NA

Closes #23002 from mgaido91/SPARK-26003.

Authored-by: Marco Gaido <marcogaido91@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit 8d7dbde914e1570a03008b0474ec647cbcb4c96e)

Cloudera ID: CDH-79190

Change-Id: I5a07cb483365ab92e747632ffb335dd6b678c774