CDH-59799. Reset the target num execs after killing
This is a backport of SPARK-21834, but a re-implementation, rather than
a cherry-pick as there are other intervening changes. Original commit
msg:

killExecutor api currently does not allow killing an executor
without updating the total number of executors needed. In case of
dynamic allocation is turned on and the allocator tries to kill an
executor, the scheduler reduces the total number of executors needed (
see
https://github.com/apache/spark/blob/master/core/src/main/scala/org/apache/spark/scheduler/cluster/CoarseGrainedSchedulerBackend.scala#L635)
which is incorrect because the allocator already takes care of setting
the required number of executors itself.

==C5_APPROVED_BUGFIX==